Fix secondary-specific tests
Many test environments are setup with primary and secondary in one.
This is not usually a problem, but in this case, these hooks were
running even though current node was stubbed as a secondary, and since
these particular tests were tagged `:sidekiq_inline`, they were running
`after_verifiable_update` which is not normally possible to trigger on a
secondary. And so the VerificationWorker ran, in the context of a
secondary. Since registry records didn't exist, they were created. The
test setup went on to create duplicate registry records (apparently we
have not enforced unique foreign keys), and anyway, the test setup was
incorrect, and so some tests failed.