Enforce :read_commit_status on CiJob
This applies the :read_commit_status policy to the `CiJob` GraphQL
entities, which are `CommitStatus` objects.

This policy guards against information exposure, and we can apply this
now that `CiJob.needs` has the correct type.

A test is added to ensure we read all fields of `CiBuildNeed` during
tests, to prevent regressions.