Document duplicate-handling of generic packages
The package file model in the database does not rely on the the package
name, version, or filename as its unique key.

As a result, it permits uploading many packages under the same entry,
while retrieving the most recently added entry when fetched back.

The UI also shows all of the separately stored entries.