Instrument CI template usage across projects
Uses Redis HLL counters to track the number of distinct projects that
run pipelines with include:template for a few templates of interest.

The tracking is behind a common feature flag so that it can all be
turned off at once if there are performance concerns.

See https://gitlab.com/gitlab-org/gitlab/-/issues/219457