Expand on use of separate name attributes in LDAP
The `name` attribute in LDAP fetches the user's visible name in profiles
but often the ideal value for this is split between two different fields
that are configurable via  the use of `first_name` and `last_name`.

The docs were not very clear on how to force use of the latter (split)
configuration over `name`, doable by intentionally setting it to
something non-existent.