Empty pattern field disabled cleanup work <13.6
In older versions of GitLab, the form for adding cleanup policies showed
a `.*` text-box placeholder text for the necessary
field `Remove tags matching` that mislead users into thinking it was
a default value. When submitted, the placeholder did not default to its
shown value and caused a `nil` value to be stored.

A `nil` value for the remove tags pattern effectively
meant no tags would get removed when the policy ran.

This has been fixed in 13.6 (placeholder also acts as default)
and 13.7 (revamp of UI).