Clarify changelog guidance documentation
Previously the changelog guidance mixed a variety of terminology and
structure, and placed the most important instructions in the middle
of the list. This attempts to place the most important instructions
at the top of the list standardize the structure of the list, and
standardize the terminology used. The intent is to make this guidance
faster and easier to take conclusions from.