Ensure MergeRequestsFinder uses correct Params
Ruby does some weird things when dynamically loading classes. It seems
in some cases it did not load `MergeRequestsFinder::Params` when it's
calling `#params_class`. In that case the superclass method is called
instead.

We discovered this issue in `IssuesFinder` before, and it was addressed
in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/41835.
This change applies the same change to `MergeRequestsFinder`.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/296939