Correctly handle Gitaly being unavailable in more locations
If `GRPC::Unavailable` or `Gitlab::Git::CommandError` is raised in code
outside of a controller action, such as an `around_action`, it can be
executed after a render call inside the controller has completed. When
this happens, ActionController detects it and raises a very unhelpful
`AbstractController::DoubleRenderError`, swallowing the original error.

We can detect this condition inside our `rescue_from` code and defer to
the usual Rails error-handling path (e.g., send the error to Sentry).
This will be easier to track down than the double-render error.