Prevent editing commits if signed commits required
On https://gitlab.com/groups/gitlab-org/-/epics/4983, we want to allow
editing commit messages on unprotected branches. Until a better
solution is found, we can't edit commit messages that are signed, so if
"Required signed commits" option is enabled, this setting shouldn't be
enabled.

This commits prevents for editing commit setting to be enabled if signed
commits is required.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/292493