Refactor code to change needs dict instead of Vue method
This change allow the needs dict to be correct and better
reflect a list of needs higher up instead of mutating it
in the Vue component. This will ensure that all the needs
are properly listed and that any consumer can
use this dict as a source of truth.

This also adds a test to catch this utility method now
accounts for parallel jobs.