Fix Service Desk issue templates dropdown data
Service Desk's templates dropdown selector FE implementation is
different from issue and merge request description templates dropdown
selector and it requres an array to be passed in instead of a hash.
Service Desk also relies on templates to be present within the project
where it is being enabled and it cannot fetch templates from other
projects that are being setup at group level, thus the need to limit the
templates that are avaiable for Service Desk to the ones defined within
the project that Service Desk is being enabled within.