Proof of concept for global issuable state
This is a proof of concept that allows for the status box
to be in control of the global issuable state on the state.
Other components can import the state directly which Vue will
add watchers to and update accordingly.
It also exports a method to update the status which gets
called when the user clicks on the comment and close button
and also whenever a polling request from the widget
is returned.