Set has_external_issue_tracker to NOT NULL
This change contains migrations for
`projects.has_external_issue_tracker` to become a `NOT NULL` column that
defaults to `FALSE`.

This is possible after the PG Trigger added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/51852 that always
maintains `projects.has_external_issue_tracker` as a boolean and never a
`NULL`.

It includes a data migration to set all historical `NULL` columns to be
either `TRUE` or `FALSE`.

The data migration also fixes the historical bad data for
`has_external_issue_tracker`:
https://gitlab.com/gitlab-org/gitlab/-/issues/273574.

https://gitlab.com/gitlab-org/gitlab/-/issues/290715.