Migrate EE MrWidgetOptions spec to @vue/test-utils
This is about as minimal a migration as is possible. There are a lot of
issues with the existing suite that have intentionally not be fixed,
e.g., mutating props directly; making assertions against component
internals; inconsistent use of async/done/promise chains. Fixing these
problems is well outside the scope of this, and can be addressed later.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/299085.