Packages to use composition, not inheritance
GraphQL does not have inheritance, so it is an anti-pattern to use
it to model polymorphism in our models.

The correct approach is to model polymorphism as implementation of
interfaces. Here there is a single `PackageType` interface, and
then different implementations of that.

Since the package types only differ in their metadata, it might be a
better idea to move the polymorphism down to the metadata fields.

One thing that had to change was the `Query.package_details` field,
which was previously `Query.package_composer_details`. This must change
since the implementation of the lookup does not perform any type
checking, and thus we cannot type the return value as a composer
package. This would be an illegal and ill-typed down-cast.

We mark the composer type as an orphan since it is only refered to
as implementing `PackageType`.