Fix diff sorting when 2 files have the same name
This to avoid an infinite loop when sorting diff files.

Before this change we try to compare identical files (e.g when
a file gets converted to symlink).

To fix it, we only try to compare unidentical files. Identical
file comparison should return 0 so `#sort` won't try to order
them.