Remove geo_framework_verification feature flag
So we don't need to keep disabling it everywhere in tests. In the next
commit, I am adding functionality which only works with replicators
which have verification enabled, so without this commit, I would have to
disable this feature flag in even more tests. Since this feature flag
doesn't guard anything which even remotely works, let's just remove it.

Since we are replacing its check with false, this commit does *not*
release anything.