Created the creation form for compliance frameworks
This form contains all the processes which are necessary to create
a compliance framework. This form makes use of the shared form to show
the necessary fields

This form saves the compliance framework via GraphQL. On success it
redirects to the general settings page. On failure, it shows an error
alert