Created the update form for compliance frameworks
This form contains all the processes which are necessary to update
a compliance framework. This form makes use of the shared form to show
the necessary fields

This form gets the compliance framework via GraphQL using the ID. On
success it populates the fields, on failure it shows an error alert.

This form saves the compliance framework via GraphQL. On success it
redirects to the general settings page. On failure, it shows an error
alert