Support auto-batching background migrations
Introduce a new background migration queueing system that tracks the
progress of each migration in the database. This allows the application
to queue background jobs per batch over time, rather than in a single
call when the migration is initially run. It also gives better control
over the execution, tracking job staus and allowing configuration
options like batch size to be adjusted on the fly.