Add 15s timeout to getaddrinfo
The URL blocker checks for resolvable addresses of the URLs it receives.

In some cases, bad/stale URLs do not resolve in a finite time, and lead
to a hang of the call for over the rack/puma timeout levels.

The `Addrinfo.getaddrinfo` call supports passing a timeout since
Ruby 2.7.0, and this change applies a 15 seconds
timeout to the operation.