Section on overriding rubocop version
A failure occurs if recent versions of Ruby are used with Code Quality
jobs.

They fail because the default rubocop version in use does not support
very recent Ruby releases (in the past 2-3 years).

An override of configuration is required to make it work in this
situation.

This change adds a troubleshooting section that specifically explains
how to override the rubocop version to address the job errors.

Related to issue:
https://gitlab.com/gitlab-org/ci-cd/codequality/-/issues/28