Add a test forcing DNS timeout
Attempts adding a test to lower the timeout to unreasonable levels
and expect a DNS lookup failure to occur.

This test may not work as expected on platforms other than Linux due
to its reliance on the `getaddrinfo_a` call existing.

Also fixed are the DNS helpers (yet again) to only look for the
timeout params