Draft: Set default MR title/description to first multi-line commit
Closes https://gitlab.com/gitlab-org/gitlab/-/issues/300479

When creating a merge request, the behaviour of the prefilled title and
discription is different than the behaviour of squashed commit messages.

For squashed commit messages, the behaviour is described in the
[docs](https://docs.gitlab.com/ee/user/project/merge_requests/squash_and_merge.html#overview)
as

> Taken from the first multi-line commit message in the merge.

However, for merge requests, the current behavior is to use the branch
name.  This commit sets the prefilled MR title/description to the first
multi-line commit in the MR, if applicable.  Otherwise falling back to
the previous behaviour.

TODO:
fix one test
update docs
generate changelog entry

WIP: QUESTIONS FOR REVIEWER

I am having trouble adding `:gitaly_commit?` to the RSPec `double`
object.  Assistance with that would be appreciated