Add migration to fix a merge_requests index
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/52116 we
introduced a new index used when generating changelogs using the GitLab
API. This migration failed in staging and production, but due to
https://github.com/lfittl/pg_query/issues/184 the original error is
masked, and an INVALID index is left behind.

This commit turns the old migration into a noop, and creates a new
migration. This new migration catches any CREATE INDEX errors and raises
a regular error, preventing the pg_query bug from surfacing. It also
removes the existing index if it's marked as INVALID. In addition, we
disable statement timeouts when creating the index; something we forgot
to do in the previous migration.