Replace circular loading icons with skeleton loaders
Unfortunately, the skeleton loader is basically just the
scanning effect across the internal SVG.

If you don't provide an SVG, you get a default one, but
to get the more custom feel, you have to provide your
own SVG.

To that end, this adds three custom loader components
which provide the relevant custom SVGs.

Similarly, the `skeleton-loader` component doesn't
provide good handles to manage to size / scaling of
the internal SVG elements, so the `merge_request`
CSS is updated to have custom styles when it's in
a loading state so the loaders aren't cut off, etc.

Finally, because there's no structure to the Diffs app 
when no content is loaded, we also add a grid layout 
to the loading state so the three loaders place 
themselves in the right spots.